+++ packages/ui/src/watch/header/Header.stories.tsx
@@ -1,0 +1,17 @@
+import type { Meta, StoryObj } from '@storybook/react';
+import { Header } from './index';
+
+const meta: Meta<typeof Header> = {
+  title: 'Watch/Header',
+  component: Header,
+  parameters: {
+    layout: 'centered',
+  },
+  tags: ['autodocs'],
+};
+
+export default meta;
+type Story = StoryObj<typeof Header>;
+
+// Basic story with default props
+export const Default: Story = {};
