+++ packages/ui/src/watch/index.ts
@@ -1,3 +1,6 @@
-import Homepage from './Homepage';
+import { HomeContainer } from './home/container';
+import { Header } from './header';
+import { SettingsPanel } from './home/settings';
+import { VideosContainer } from './home/videos-container';
 
-export { Homepage };
+export { HomeContainer, Header, SettingsPanel, VideosContainer };
